+++ Wacs.Core/Wacs.Core/Instructions/Memory/InstI32Store.cs
@@ -56,7 +56,7 @@ namespace Wacs.Core.Instructions.Memory
             if ((ulong)ea > (ulong)mem.ByteLength || (ulong)mem.ByteLength - (ulong)ea < (ulong)WidthTByteSize)
                 throw new TrapException($"Instruction {Op.GetMnemonic()} failed. Memory pointer out of bounds.");
             //13,14,15
-            Span<byte> bs = mem.AsSpan((int)ea, WidthTByteSize);
+            Span<byte> bs = mem.AsSpan((nuint)ea, WidthTByteSize);
 
 #if NET8_0_OR_GREATER
             MemoryMarshal.Write(bs, in cU32);
@@ -99,7 +99,7 @@ namespace Wacs.Core.Instructions.Memory
             if ((ulong)ea > (ulong)mem.ByteLength || (ulong)mem.ByteLength - (ulong)ea < (ulong)WidthTByteSize)
                 throw new TrapException($"Instruction {Op.GetMnemonic()} failed. Memory pointer out of bounds.");
             // Single-byte store via mode-aware AsSpan.
-            mem.AsSpan((int)ea, 1)[0] = (byte)(0xFF & cU32);
+            mem.AsSpan((nuint)ea, 1)[0] = (byte)(0xFF & cU32);
         }
     }
     
@@ -136,7 +136,7 @@ namespace Wacs.Core.Instructions.Memory
             if ((ulong)ea > (ulong)mem.ByteLength || (ulong)mem.ByteLength - (ulong)ea < (ulong)WidthTByteSize)
                 throw new TrapException($"Instruction {Op.GetMnemonic()} failed. Memory pointer out of bounds.");
             //13,14,15
-            Span<byte> bs = mem.AsSpan((int)ea, WidthTByteSize);
+            Span<byte> bs = mem.AsSpan((nuint)ea, WidthTByteSize);
             
             ushort cI16 = (ushort)cU32;
 #if NET8_0_OR_GREATER
