+++ Wacs.Core/Wacs.Core/Instructions/Memory/InstI64Store.cs
@@ -56,7 +56,7 @@ namespace Wacs.Core.Instructions.Memory
             if ((ulong)ea > (ulong)mem.ByteLength || (ulong)mem.ByteLength - (ulong)ea < (ulong)WidthTByteSize)
                 throw new TrapException($"Instruction {Op.GetMnemonic()} failed. Memory pointer out of bounds.");
             //13,14,15
-            Span<byte> bs = mem.AsSpan((int)ea, WidthTByteSize);
+            Span<byte> bs = mem.AsSpan((nuint)ea, WidthTByteSize);
             
 #if NET8_0_OR_GREATER
             MemoryMarshal.Write(bs, in cU64);
@@ -99,7 +99,7 @@ namespace Wacs.Core.Instructions.Memory
             if ((ulong)ea > (ulong)mem.ByteLength || (ulong)mem.ByteLength - (ulong)ea < (ulong)WidthTByteSize)
                 throw new TrapException($"Instruction {Op.GetMnemonic()} failed. Memory pointer out of bounds.");
             //13,14,15
-            mem.AsSpan((int)ea, 1)[0] = (byte)(0xFF & cU64);
+            mem.AsSpan((nuint)ea, 1)[0] = (byte)(0xFF & cU64);
         }
     }
     
@@ -136,7 +136,7 @@ namespace Wacs.Core.Instructions.Memory
             if ((ulong)ea > (ulong)mem.ByteLength || (ulong)mem.ByteLength - (ulong)ea < (ulong)WidthTByteSize)
                 throw new TrapException($"Instruction {Op.GetMnemonic()} failed. Memory pointer out of bounds.");
             //13,14,15
-            Span<byte> bs = mem.AsSpan((int)ea, WidthTByteSize);
+            Span<byte> bs = mem.AsSpan((nuint)ea, WidthTByteSize);
             
             ushort cI16 = (ushort)cU64;
 #if NET8_0_OR_GREATER
@@ -180,7 +180,7 @@ namespace Wacs.Core.Instructions.Memory
             if ((ulong)ea > (ulong)mem.ByteLength || (ulong)mem.ByteLength - (ulong)ea < (ulong)WidthTByteSize)
                 throw new TrapException($"Instruction {Op.GetMnemonic()} failed. Memory pointer out of bounds.");
             //13,14,15
-            Span<byte> bs = mem.AsSpan((int)ea, WidthTByteSize);
+            Span<byte> bs = mem.AsSpan((nuint)ea, WidthTByteSize);
             
             uint cI32 = (uint)cU64;
 #if NET8_0_OR_GREATER
